Unspecified charges
Charges of unspecified tincture now correctly draw as the next named
tincture (eg "azure, a bend between two mullets or" now draws the bend
or as well as the mullets). This was acieved through a massive
re-architecting of the parsing functions. They all now take a tree
(which now has a defined active node) and a success array, and return
the new tree.