Render semy charges; new rendering system
We now have a function that draws semy charges on arbitrary elements,
but it is not used yet. We also have the beginnings of a new method of
drawing charges etc, which relies on the DOM less and so should be
quicker (and able eventually to be moved into a webworker).